% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAGR.R
\name{CAGR}
\alias{CAGR}
\title{Compute CAGR(Compound Annual Growth Rate)}
\usage{
CAGR(data.1, data.n, n)
}
\arguments{
\item{data.1}{Data of the first year}

\item{data.n}{Data of the last year}

\item{n}{Number of years}
}
\value{
CAGR and between years values
}
\description{
Compute CAGR(Compound Annual Growth Rate)
}
\examples{
c.cagr<-CAGR(100, 189, 5)
}
\references{
Bardhan, D., Singh, S.R.K., Raut, A.A.and Athare, T.R. (2022). Livestock in Madhya Pradesh and Chhattisgarh: An Analysis for Some Policy Implications. Agricultural Science Digest. DOI:10.18805/ag.D-5418.
}
