% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAGR.R
\name{data.first}
\alias{data.first}
\title{Computing First Year data}
\usage{
data.first(data.n, r, n)
}
\arguments{
\item{data.n}{data of the last year}

\item{r}{CAGR}

\item{n}{number of years}
}
\value{
First year data and between years values
}
\description{
Computing first year data
}
\examples{
d.first<-data.first(189, 13.57751, 5)
}
\references{
Bardhan, D., Singh, S.R.K., Raut, A.A.and Athare, T.R. (2022). Livestock in Madhya Pradesh and Chhattisgarh: An Analysis for Some Policy Implications. Agricultural Science Digest. DOI:10.18805/ag.D-5418.
}
