% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAGR.R
\name{data.last}
\alias{data.last}
\title{Computing Last Year data}
\usage{
data.last(data.1, r, n)
}
\arguments{
\item{data.1}{data of the first year}

\item{r}{CAGR}

\item{n}{number of years}
}
\value{
Last year data and between years values
}
\description{
Computing last year data
}
\examples{
d.last<-data.last(100, 13.57751, 5)
}
\references{
Bardhan, D., Singh, S.R.K., Raut, A.A.and Athare, T.R. (2022). Livestock in Madhya Pradesh and Chhattisgarh: An Analysis for Some Policy Implications. Agricultural Science Digest. DOI:10.18805/ag.D-5418.
}
