% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAGR.R
\name{LLTM}
\alias{LLTM}
\title{Log-linear Trend Model}
\usage{
LLTM(data)
}
\arguments{
\item{data}{Time series data set}
}
\value{
Growth rate and model summary
}
\description{
Estimating the growth rate using the log-linear trend model
}
\examples{
data<- c(100,105,110,118,116,120,130)
g.rate <- LLTM(data)
}
\references{
Bardhan, D., Singh, S.R.K., Raut, A.A.and Athare, T.R. (2022). Livestock in Madhya Pradesh and Chhattisgarh: An Analysis for Some Policy Implications. Agricultural Science Digest. DOI:10.18805/ag.D-5418.
}
