% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAGR.R
\name{n.years}
\alias{n.years}
\title{Computing Number of Years}
\usage{
n.years(data.1, data.n, r)
}
\arguments{
\item{data.1}{Data of the first year}

\item{data.n}{Data of the last year}

\item{r}{CAGR}
}
\value{
Number of years and between years values
}
\description{
Computing number of years
}
\examples{
n.yrs<-n.years(100, 189, 13.57751)
}
\references{
Bardhan, D., Singh, S.R.K., Raut, A.A.and Athare, T.R. (2022). Livestock in Madhya Pradesh and Chhattisgarh: An Analysis for Some Policy Implications. Agricultural Science Digest. DOI:10.18805/ag.D-5418.
}
